Reverted some changes for the network unknowns and residuals naming
Naming Convention works for both root finder and optimze(except ground segment). The Boeing 787 can be used to test this
THe secondary solver to solve the network can only be used with the root solver.
Removed the combined solver flag. SInce if the network solver is none is automatically solves it togther
Will work on fixing the battery now